requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.0
psycopg2-binary==2.9.5
scrapy==2.8.0
Flask-WTF==1.0.0
WTForms==3.0.0
Flask-Migrate
